backend/requirements.txt
fastapi
uvicorn
krutrim-cloud
python-dotenv
pydantic
